Lesson 10: part 1
Seeing as I don't feel like the hover tooltip is really necessary at the
moment with the title function doing the same, I added the text comments
when you click on a text field. I may change this, I was having problems
with the fade in portion and a few others. If I resolve those I'll
upload those.